refactor(plugins): Move KSP compiler to separate project
Move the KSP compiler to a separate project to better separate the
plugin API from the compiler plugin.
