+++ plugins/advisors/vulnerable-code/build.gradle.kts
@@ -34,7 +34,7 @@ dependencies {
     implementation(projects.utils.ortUtils)
 
     ksp(projects.advisor)
-    ksp(projects.plugins.api)
+    ksp(projects.plugins.compiler)
 
     testImplementation(libs.wiremock)
 }
