+++ plugins/compiler/src/main/resources/META-INF/services/com.google.devtools.ksp.processing.SymbolProcessorProvider
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.compiler.PluginProcessorProvider
